refactor(rate_of_closure): split glossary data modules to honor the 500-LOC budget
glossary.py / glossary.ts keep the public surface (GLOSSARY,
FIELD_TO_TERM, search) and merge the split entry-data modules
(glossary_entries_a_l / _l_z, glossaryEntriesAL/LZ.ts, shared type
modules). No behavior change; all glossary tests and parity pins pass.
