+++ src/rate_of_closure/glossary_entries_l_z.py
@@ -1,0 +1,243 @@
+"""Glossary entry data (lever_arm..uniform_distribution) — see glossary.
+
+Split from the glossary module to honor the 500-LOC file budget; the
+public surface (``GLOSSARY``, ``FIELD_TO_TERM``, ``search_terms``)
+lives in :mod:`rate_of_closure.glossary`, which merges these dicts and
+validates them.
+"""
+
+from __future__ import annotations
+
+from ._contracts import require
+from .glossary_types import GlossaryEntry
+
+__all__ = ["ENTRIES"]
+
+
+def _entry(term: str, definition: str) -> GlossaryEntry:
+    require(bool(term.strip()), "glossary term must be non-empty", term)
+    require(
+        len(definition.strip()) >= 60,
+        "glossary definitions must be substantive",
+        term,
+    )
+    return GlossaryEntry(term=term, definition=definition)
+
+
+ENTRIES: dict[str, GlossaryEntry] = {
+    "lever_arm": _entry(
+        "Lever Arm",
+        "The vector r from the reference point (GC or CG) to the struck "
+        "point. It converts rotation into extra point velocity (ω × r) "
+        "in the closure model and impulse into recoil torque (r × Jn) in "
+        "the impact model.",
+    ),
+    "lie_angle": _entry(
+        "Lie Angle",
+        "The angle between the shaft and the ground plane at impact. It "
+        "sets how the shaft-axis (HTV) and swing-plane (SPV) rotation "
+        "rates combine into face closure: CCV = HTV sin(lie) + SPV "
+        "cos(lie) (Cheetham 2014 reconciliation).",
+    ),
+    "lift": _entry(
+        "Lift Coefficient (Cl)",
+        "The dimensionless coefficient of the aerodynamic force "
+        "perpendicular to the ball's motion, generated by backspin (the "
+        "Magnus effect). Literature models express Cl as a function of "
+        "spin ratio, capped at a physical maximum "
+        "(swing_sim.flight.models).",
+    ),
+    "magnus_force": _entry(
+        "Magnus Force",
+        "The aerodynamic force on a spinning ball, perpendicular to both "
+        "the velocity and the spin axis: backspin lifts the ball, a "
+        "tilted spin axis curves it sideways. It enters the flight EOM "
+        "through the lift term (swing_sim.flight; Penner 2003).",
+    ),
+    "mass_matrix": _entry(
+        "Mass Matrix",
+        "The configuration-dependent 2x2 (or 3x3) inertia matrix M(θ) of "
+        "the pendulum equations M(θ)·α + C(θ, ω) + G(θ) + D(ω) = 0; its "
+        "off-diagonal terms couple the links through the elbow/wrist "
+        "angle (swing_sim.reference.mass_matrix).",
+    ),
+    "moi": _entry(
+        "Moment of Inertia (MOI)",
+        "A body's resistance to angular acceleration about an axis. The "
+        "clubhead MOI (scalar or full 3x3 tensor) sets how much an "
+        "off-center impulse twists the head instead of launching the "
+        "ball — the club-side term of the effective mass "
+        "(swing_sim.impact.models).",
+    ),
+    "moi_tensor": _entry(
+        "MOI Tensor",
+        "The full 3x3 inertia tensor I of the clubhead. The exact "
+        "off-center effective mass uses the triple-product form "
+        "(r x n)^T I^-1 (r x n); a diagonal I·eye(3) reproduces the "
+        "scalar-MOI fallback 1/m + |r|²/I exactly "
+        "(swing_sim.impact.models derivation).",
+    ),
+    "monte_carlo": _entry(
+        "Monte Carlo Simulation",
+        "Running the simulation many times with randomized inputs drawn "
+        "from per-variable noise distributions, then reading dispersion "
+        "statistics off the output sample. The Variation tab's seeded "
+        "engine makes runs exactly reproducible "
+        "(swing_sim.variation.engine).",
+    ),
+    "noise_spec": _entry(
+        "Noise Specification (NoiseSpec)",
+        "The per-variable description of how an input varies in a "
+        "variation study: distribution family (normal, uniform, or "
+        "triangular), additive scale, and optional clip truncation "
+        "(swing_sim.variation.spec).",
+    ),
+    "normal_distribution": _entry(
+        "Normal Distribution",
+        "The bell-curve distribution parameterized by mean and standard "
+        "deviation — the default noise family for delivery variables in "
+        "variation studies, matching how measurement scatter is usually "
+        "reported (swing_sim.variation registry guidance).",
+    ),
+    "one_at_a_time_sensitivity": _entry(
+        "One-at-a-Time (OAT) Sensitivity",
+        "A sensitivity method that re-runs the study with only one input "
+        "varying at a time, using paired random draws, to attribute "
+        "output variance to individual inputs "
+        "(swing_sim.variation.analysis).",
+    ),
+    "pitch": _entry(
+        "Screw Pitch",
+        "The ratio of translation along the instantaneous screw axis to "
+        "rotation about it. A pure rotation has zero pitch; the "
+        "clubhead's near-impact motion has a small pitch, which is why "
+        "the screw-axis picture works (screw theory; AffineDrift "
+        "rotation review).",
+    ),
+    "plane_inclination": _entry(
+        "Swing-Plane Inclination",
+        "The orientation of the pendulum swing plane in space — yaw, "
+        "side tilt, and forward tilt. Gravity is projected into the "
+        "plane (g_inplane = R^T (0, 0, -g)), so steeper planes feel more "
+        "in-plane gravity (swing_sim.reference.in_plane_gravity).",
+    ),
+    "r_isa": _entry(
+        "Distance to the Screw Axis (R_ISA)",
+        "The distance v/ω from the clubhead to the instantaneous screw "
+        "axis. Closure per foot equals 1 / R_ISA, and the path gap "
+        "between two reference points separated by d is d / R_ISA — "
+        "independent of clubhead speed (AffineDrift closure derivation).",
+    ),
+    "roll": _entry(
+        "Roll",
+        "The vertical (crown-sole) curvature of a wood's face. It adds "
+        "loft to high strikes and removes it from low ones, partially "
+        "compensating the gear effect's backspin change (club-design "
+        "literature; rate_of_closure.club face model).",
+    ),
+    "screw_axis": _entry(
+        "Instantaneous Screw Axis (ISA)",
+        "The unique line about which a rigid body's motion at one "
+        "instant is a rotation plus a slide along that same line "
+        "(Chasles' theorem). The clubhead sweeps about it near impact; "
+        "its distance is R_ISA (screw theory; AffineDrift rotation "
+        "review).",
+    ),
+    "seed": _entry(
+        "Random Seed",
+        "The integer that initializes the pseudo-random number generator "
+        "of a variation study. The engine derives a per-variable stream "
+        "from [seed, crc32(key)], so the same plan and seed always "
+        "reproduce the same dataset (swing_sim.variation.engine).",
+    ),
+    "sensitivity_analysis": _entry(
+        "Sensitivity Analysis",
+        "Quantifying which inputs drive which outputs. The Variation tab "
+        "combines one-at-a-time reruns (local attribution) with Spearman "
+        "rank correlation on the full dataset (a cheap global "
+        "cross-check) (swing_sim.variation.analysis).",
+    ),
+    "smash_factor": _entry(
+        "Smash Factor",
+        "Ball speed divided by clubhead speed — the standard efficiency "
+        "measure of an impact. A well-struck driver reaches about "
+        "1.48-1.50; off-center hits lose smash through the reduced "
+        "effective mass (launch-monitor norms).",
+    ),
+    "spearman": _entry(
+        "Spearman Rank Correlation",
+        "A correlation computed on ranks rather than raw values, so it "
+        "captures any monotonic input-output relationship without "
+        "assuming linearity. Used as the global sensitivity cross-check "
+        "in variation studies (swing_sim.variation.analysis).",
+    ),
+    "spin_axis_tilt": _entry(
+        "Spin-Axis Tilt",
+        "The tilt of the ball's spin axis away from horizontal (+ = "
+        "fade side, right of target). Set by the face-minus-path "
+        "difference through the D-plane; it converts backspin into "
+        "sideways curvature (TrackMan D-plane literature).",
+    ),
+    "spin_decay": _entry(
+        "Spin Decay",
+        "The gradual loss of spin rate during flight from aerodynamic "
+        "torque, modeled as an exponential decay in the "
+        "MacDonald-Hanzely and constant-coefficient flight models "
+        "(swing_sim.flight.models).",
+    ),
+    "spin_loft": _entry(
+        "Spin Loft",
+        "The 3-D angle between the delivered face normal and the club "
+        "path vector: spin_loft = arccos(v̂ · n̂). It sets how much of "
+        "the impact goes into spin instead of speed "
+        "(swing_sim.impact.delivery; TrackMan conventions).",
+    ),
+    "spin_rate": _entry(
+        "Spin Rate",
+        "The ball's total rotation rate in rpm — the friction impulse of "
+        "the impact solve (capped at the 2/7 rolling limit) plus "
+        "gear-effect spin. Driver band roughly 2,000-3,500 rpm "
+        "(swing_sim.impact; launch-monitor norms).",
+    ),
+    "spv": _entry(
+        "Swing-Plane Velocity (SPV)",
+        "The clubhead's angular velocity about the swing-plane normal — "
+        "the in-plane rotation of the swing arc. Together with HTV it "
+        "assembles the full angular velocity vector (Cheetham 2014 "
+        "3-D motion studies).",
+    ),
+    "time_to_square": _entry(
+        "Time to Square",
+        "How long before impact the face was one degree open at the "
+        "current closure rate — about half a millisecond at tour rates, "
+        "the classic framing of release-timing tolerance (closure-rate "
+        "dossier).",
+    ),
+    "triangular_distribution": _entry(
+        "Triangular Distribution",
+        "A bounded distribution rising linearly to a peak and falling "
+        "back — a practical choice in variation studies when only a "
+        "min / most-likely / max estimate is available "
+        "(swing_sim.variation.spec).",
+    ),
+    "triple_pendulum": _entry(
+        "Triple Pendulum Swing Model",
+        "A three-link extension of the double pendulum (torso-arms-club) "
+        "solved with the same mass-matrix formalism in a planar frame. "
+        "Available as a swing source in the Simulation tab "
+        "(rate_of_closure.simulation.sources).",
+    ),
+    "twist": _entry(
+        "Twist",
+        "A rigid body's instantaneous motion state — angular velocity "
+        "plus the linear velocity of a reference point. The twist "
+        "relation v_P = v_ref + ω × r gives every point's velocity from "
+        "one twist (screw theory; the core of the closure model).",
+    ),
+    "uniform_distribution": _entry(
+        "Uniform Distribution",
+        "A distribution giving every value in a bounded interval equal "
+        "probability — used in variation studies when only hard limits, "
+        "not a central tendency, are known (swing_sim.variation.spec).",
+    ),
+}
