+++ src/rate_of_closure/web/src/model/glossaryEntriesLZ.ts
@@ -1,0 +1,245 @@
+/**
+ * Glossary entry data (lever_arm through uniform_distribution) — see glossary.ts.
+ *
+ * Split from the glossary module to honor the 500-LOC file budget.
+ * Generated from `src/rate_of_closure/glossary.py`.
+ */
+
+import type { GlossaryEntry } from "./glossaryTypes";
+
+export const ENTRIES: Record<string, GlossaryEntry> = {
+  lever_arm: {
+    term: "Lever Arm",
+    definition:
+      "The vector r from the reference point (GC or CG) to the struck " +
+      "point. It converts rotation into extra point velocity (\u03c9 \u00d7 r) in the " +
+      "closure model and impulse into recoil torque (r \u00d7 Jn) in the impact " +
+      "model.",
+  },
+  lie_angle: {
+    term: "Lie Angle",
+    definition:
+      "The angle between the shaft and the ground plane at impact. It sets " +
+      "how the shaft-axis (HTV) and swing-plane (SPV) rotation rates " +
+      "combine into face closure: CCV = HTV sin(lie) + SPV cos(lie) " +
+      "(Cheetham 2014 reconciliation).",
+  },
+  lift: {
+    term: "Lift Coefficient (Cl)",
+    definition:
+      "The dimensionless coefficient of the aerodynamic force perpendicular " +
+      "to the ball's motion, generated by backspin (the Magnus effect). " +
+      "Literature models express Cl as a function of spin ratio, capped at " +
+      "a physical maximum (swing_sim.flight.models).",
+  },
+  magnus_force: {
+    term: "Magnus Force",
+    definition:
+      "The aerodynamic force on a spinning ball, perpendicular to both the " +
+      "velocity and the spin axis: backspin lifts the ball, a tilted spin " +
+      "axis curves it sideways. It enters the flight EOM through the lift " +
+      "term (swing_sim.flight; Penner 2003).",
+  },
+  mass_matrix: {
+    term: "Mass Matrix",
+    definition:
+      "The configuration-dependent 2x2 (or 3x3) inertia matrix M(\u03b8) of the " +
+      "pendulum equations M(\u03b8)\u00b7\u03b1 + C(\u03b8, \u03c9) + G(\u03b8) + D(\u03c9) = 0; its " +
+      "off-diagonal terms couple the links through the elbow/wrist angle " +
+      "(swing_sim.reference.mass_matrix).",
+  },
+  moi: {
+    term: "Moment of Inertia (MOI)",
+    definition:
+      "A body's resistance to angular acceleration about an axis. The " +
+      "clubhead MOI (scalar or full 3x3 tensor) sets how much an off-center " +
+      "impulse twists the head instead of launching the ball \u2014 the " +
+      "club-side term of the effective mass (swing_sim.impact.models).",
+  },
+  moi_tensor: {
+    term: "MOI Tensor",
+    definition:
+      "The full 3x3 inertia tensor I of the clubhead. The exact off-center " +
+      "effective mass uses the triple-product form (r x n)^T I^-1 (r x n); " +
+      "a diagonal I\u00b7eye(3) reproduces the scalar-MOI fallback 1/m + |r|\u00b2/I " +
+      "exactly (swing_sim.impact.models derivation).",
+  },
+  monte_carlo: {
+    term: "Monte Carlo Simulation",
+    definition:
+      "Running the simulation many times with randomized inputs drawn from " +
+      "per-variable noise distributions, then reading dispersion statistics " +
+      "off the output sample. The Variation tab's seeded engine makes runs " +
+      "exactly reproducible (swing_sim.variation.engine).",
+  },
+  noise_spec: {
+    term: "Noise Specification (NoiseSpec)",
+    definition:
+      "The per-variable description of how an input varies in a variation " +
+      "study: distribution family (normal, uniform, or triangular), " +
+      "additive scale, and optional clip truncation " +
+      "(swing_sim.variation.spec).",
+  },
+  normal_distribution: {
+    term: "Normal Distribution",
+    definition:
+      "The bell-curve distribution parameterized by mean and standard " +
+      "deviation \u2014 the default noise family for delivery variables in " +
+      "variation studies, matching how measurement scatter is usually " +
+      "reported (swing_sim.variation registry guidance).",
+  },
+  one_at_a_time_sensitivity: {
+    term: "One-at-a-Time (OAT) Sensitivity",
+    definition:
+      "A sensitivity method that re-runs the study with only one input " +
+      "varying at a time, using paired random draws, to attribute output " +
+      "variance to individual inputs (swing_sim.variation.analysis).",
+  },
+  pitch: {
+    term: "Screw Pitch",
+    definition:
+      "The ratio of translation along the instantaneous screw axis to " +
+      "rotation about it. A pure rotation has zero pitch; the clubhead's " +
+      "near-impact motion has a small pitch, which is why the screw-axis " +
+      "picture works (screw theory; AffineDrift rotation review).",
+  },
+  plane_inclination: {
+    term: "Swing-Plane Inclination",
+    definition:
+      "The orientation of the pendulum swing plane in space \u2014 yaw, side " +
+      "tilt, and forward tilt. Gravity is projected into the plane " +
+      "(g_inplane = R^T (0, 0, -g)), so steeper planes feel more in-plane " +
+      "gravity (swing_sim.reference.in_plane_gravity).",
+  },
+  r_isa: {
+    term: "Distance to the Screw Axis (R_ISA)",
+    definition:
+      "The distance v/\u03c9 from the clubhead to the instantaneous screw axis. " +
+      "Closure per foot equals 1 / R_ISA, and the path gap between two " +
+      "reference points separated by d is d / R_ISA \u2014 independent of " +
+      "clubhead speed (AffineDrift closure derivation).",
+  },
+  roll: {
+    term: "Roll",
+    definition:
+      "The vertical (crown-sole) curvature of a wood's face. It adds loft " +
+      "to high strikes and removes it from low ones, partially compensating " +
+      "the gear effect's backspin change (club-design literature; " +
+      "rate_of_closure.club face model).",
+  },
+  screw_axis: {
+    term: "Instantaneous Screw Axis (ISA)",
+    definition:
+      "The unique line about which a rigid body's motion at one instant is " +
+      "a rotation plus a slide along that same line (Chasles' theorem). The " +
+      "clubhead sweeps about it near impact; its distance is R_ISA (screw " +
+      "theory; AffineDrift rotation review).",
+  },
+  seed: {
+    term: "Random Seed",
+    definition:
+      "The integer that initializes the pseudo-random number generator of a " +
+      "variation study. The engine derives a per-variable stream from " +
+      "[seed, crc32(key)], so the same plan and seed always reproduce the " +
+      "same dataset (swing_sim.variation.engine).",
+  },
+  sensitivity_analysis: {
+    term: "Sensitivity Analysis",
+    definition:
+      "Quantifying which inputs drive which outputs. The Variation tab " +
+      "combines one-at-a-time reruns (local attribution) with Spearman rank " +
+      "correlation on the full dataset (a cheap global cross-check) " +
+      "(swing_sim.variation.analysis).",
+  },
+  smash_factor: {
+    term: "Smash Factor",
+    definition:
+      "Ball speed divided by clubhead speed \u2014 the standard efficiency " +
+      "measure of an impact. A well-struck driver reaches about 1.48-1.50; " +
+      "off-center hits lose smash through the reduced effective mass " +
+      "(launch-monitor norms).",
+  },
+  spearman: {
+    term: "Spearman Rank Correlation",
+    definition:
+      "A correlation computed on ranks rather than raw values, so it " +
+      "captures any monotonic input-output relationship without assuming " +
+      "linearity. Used as the global sensitivity cross-check in variation " +
+      "studies (swing_sim.variation.analysis).",
+  },
+  spin_axis_tilt: {
+    term: "Spin-Axis Tilt",
+    definition:
+      "The tilt of the ball's spin axis away from horizontal (+ = fade " +
+      "side, right of target). Set by the face-minus-path difference " +
+      "through the D-plane; it converts backspin into sideways curvature " +
+      "(TrackMan D-plane literature).",
+  },
+  spin_decay: {
+    term: "Spin Decay",
+    definition:
+      "The gradual loss of spin rate during flight from aerodynamic torque, " +
+      "modeled as an exponential decay in the MacDonald-Hanzely and " +
+      "constant-coefficient flight models (swing_sim.flight.models).",
+  },
+  spin_loft: {
+    term: "Spin Loft",
+    definition:
+      "The 3-D angle between the delivered face normal and the club path " +
+      "vector: spin_loft = arccos(v\u0302 \u00b7 n\u0302). It sets how much of the impact " +
+      "goes into spin instead of speed (swing_sim.impact.delivery; TrackMan " +
+      "conventions).",
+  },
+  spin_rate: {
+    term: "Spin Rate",
+    definition:
+      "The ball's total rotation rate in rpm \u2014 the friction impulse of the " +
+      "impact solve (capped at the 2/7 rolling limit) plus gear-effect " +
+      "spin. Driver band roughly 2,000-3,500 rpm (swing_sim.impact; " +
+      "launch-monitor norms).",
+  },
+  spv: {
+    term: "Swing-Plane Velocity (SPV)",
+    definition:
+      "The clubhead's angular velocity about the swing-plane normal \u2014 the " +
+      "in-plane rotation of the swing arc. Together with HTV it assembles " +
+      "the full angular velocity vector (Cheetham 2014 3-D motion studies).",
+  },
+  time_to_square: {
+    term: "Time to Square",
+    definition:
+      "How long before impact the face was one degree open at the current " +
+      "closure rate \u2014 about half a millisecond at tour rates, the classic " +
+      "framing of release-timing tolerance (closure-rate dossier).",
+  },
+  triangular_distribution: {
+    term: "Triangular Distribution",
+    definition:
+      "A bounded distribution rising linearly to a peak and falling back \u2014 " +
+      "a practical choice in variation studies when only a min / " +
+      "most-likely / max estimate is available (swing_sim.variation.spec).",
+  },
+  triple_pendulum: {
+    term: "Triple Pendulum Swing Model",
+    definition:
+      "A three-link extension of the double pendulum (torso-arms-club) " +
+      "solved with the same mass-matrix formalism in a planar frame. " +
+      "Available as a swing source in the Simulation tab " +
+      "(rate_of_closure.simulation.sources).",
+  },
+  twist: {
+    term: "Twist",
+    definition:
+      "A rigid body's instantaneous motion state \u2014 angular velocity plus " +
+      "the linear velocity of a reference point. The twist relation v_P = " +
+      "v_ref + \u03c9 \u00d7 r gives every point's velocity from one twist (screw " +
+      "theory; the core of the closure model).",
+  },
+  uniform_distribution: {
+    term: "Uniform Distribution",
+    definition:
+      "A distribution giving every value in a bounded interval equal " +
+      "probability \u2014 used in variation studies when only hard limits, not a " +
+      "central tendency, are known (swing_sim.variation.spec).",
+  },
+};
